initial split out of components
