+++ easy-ui-react/src/VerticalNav/types.d.ts
@@ -1,0 +1,7 @@
+import { AriaLabelingProps } from "@react-types/shared";
+import { ReactNode } from "react";
+
+export type BaseVerticalNavProps = AriaLabelingProps & {
+  children: ReactNode;
+  renderLogo?: () => ReactNode;
+};
